rename error summary
